get all links function
